:art: Move repeatable code to core
