+++ src/apps/core/views.py
@@ -1,3 +1,20 @@
-from django.shortcuts import render
+from revproxy.views import DiazoProxyView
+import json
 
-# Create your views here.
+
+class EdemProxyView(DiazoProxyView):
+    html5 = True
+
+    def dispatch(self, request, *args, **kwargs):
+        self.request = request
+
+        if request.user.is_authenticated:
+            user_data = {
+                'name': request.user.first_name,
+                'email': request.user.email,
+            }
+
+            request.META['HTTP_REMOTE_USER_DATA'] = json.dumps(user_data)
+            request.META['HTTP_AUTH_USER'] = request.user.username
+
+        return super(EdemProxyView, self).dispatch(request, *args, **kwargs)
