+++ src/apps/wikilegis/views.py
@@ -1,25 +1,7 @@
-from revproxy.views import DiazoProxyView
+from apps.core.views import EdemProxyView
 from django.conf import settings
-import json
 
 
-class WikilegisProxyView(DiazoProxyView):
+class WikilegisProxyView(EdemProxyView):
     upstream = settings.WIKILEGIS_UPSTREAM
     diazo_theme_template = 'diazo-base.html'
-    html5 = True
-
-    def dispatch(self, request, *args, **kwargs):
-        self.request = request
-
-        if request.user.is_authenticated:
-            user_data = {
-                'name': request.user.first_name,
-                'email': request.user.email,
-            }
-
-            request.META['HTTP_REMOTE_USER_DATA'] = json.dumps(user_data)
-            request.META['HTTP_AUTH_USER'] = request.user.username
-
-        return super(WikilegisProxyView, self).dispatch(
-            request, *args, **kwargs
-        )
